GH-1538: Move getFreeVars method to QueryAlgebraUtil
This change moves the general method that inspects TupleExpr for free
variables to QueryAlgebraUtils. Background is to have it in a more
convenient place.

In addition the computation of free variables for LeftJoin is improved.
