+++ tools/federation/src/main/java/org/eclipse/rdf4j/federated/optimizer/StatementGroupAndJoinOptimizer.java
@@ -8,7 +8,6 @@
 package org.eclipse.rdf4j.federated.optimizer;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.LinkedList;
@@ -20,20 +19,12 @@ import org.eclipse.rdf4j.federated.algebra.EmptyResult;
 import org.eclipse.rdf4j.federated.algebra.ExclusiveGroup;
 import org.eclipse.rdf4j.federated.algebra.ExclusiveStatement;
 import org.eclipse.rdf4j.federated.algebra.ExclusiveTupleExpr;
-import org.eclipse.rdf4j.federated.algebra.FedXService;
 import org.eclipse.rdf4j.federated.algebra.NJoin;
-import org.eclipse.rdf4j.federated.algebra.NTuple;
-import org.eclipse.rdf4j.federated.algebra.StatementTupleExpr;
 import org.eclipse.rdf4j.federated.exception.OptimizationException;
 import org.eclipse.rdf4j.federated.structures.QueryInfo;
-import org.eclipse.rdf4j.query.algebra.ArbitraryLengthPath;
-import org.eclipse.rdf4j.query.algebra.BindingSetAssignment;
-import org.eclipse.rdf4j.query.algebra.Extension;
-import org.eclipse.rdf4j.query.algebra.LeftJoin;
-import org.eclipse.rdf4j.query.algebra.Projection;
+import org.eclipse.rdf4j.federated.util.QueryAlgebraUtil;
 import org.eclipse.rdf4j.query.algebra.QueryModelNode;
 import org.eclipse.rdf4j.query.algebra.Service;
-import org.eclipse.rdf4j.query.algebra.StatementPattern;
 import org.eclipse.rdf4j.query.algebra.TupleExpr;
 import org.eclipse.rdf4j.query.algebra.helpers.AbstractQueryModelVisitor;
 import org.slf4j.Logger;
@@ -271,7 +262,7 @@ public class StatementGroupAndJoinOptimizer extends AbstractQueryModelVisitor<Op
 				}
 			}
 
-			joinVars.addAll(getFreeVars(item));
+			joinVars.addAll(QueryAlgebraUtil.getFreeVars(item));
 			if (log.isTraceEnabled())
 				log.trace("Cost of " + item.getClass().getSimpleName() + " is determined as " + minCost);
 			optimized.add(item);
@@ -283,67 +274,5 @@ public class StatementGroupAndJoinOptimizer extends AbstractQueryModelVisitor<Op
 
 	protected double estimateCost(TupleExpr tupleExpr, Set<String> joinVars) {
 		return costModel.estimateCost(tupleExpr, joinVars);
-	}
-
-	public static Collection<String> getFreeVars(TupleExpr tupleExpr) {
-		if (tupleExpr instanceof StatementTupleExpr)
-			return ((StatementTupleExpr) tupleExpr).getFreeVars();
-
-		// determine the number of free variables in a UNION or Join
-		if (tupleExpr instanceof NTuple) {
-			HashSet<String> freeVars = new HashSet<>();
-			NTuple ntuple = (NTuple) tupleExpr;
-			for (TupleExpr t : ntuple.getArgs())
-				freeVars.addAll(getFreeVars(t));
-			return freeVars;
-		}
-
-		if (tupleExpr instanceof FedXService) {
-			return ((FedXService) tupleExpr).getFreeVars();
-		}
-
-		if (tupleExpr instanceof Service) {
-			return ((Service) tupleExpr).getServiceVars();
-		}
-
-		// can happen in SERVICE nodes, if they cannot be optimized
-		if (tupleExpr instanceof StatementPattern) {
-			List<String> freeVars = new ArrayList<>();
-			StatementPattern st = (StatementPattern) tupleExpr;
-			if (st.getSubjectVar().getValue() == null)
-				freeVars.add(st.getSubjectVar().getName());
-			if (st.getPredicateVar().getValue() == null)
-				freeVars.add(st.getPredicateVar().getName());
-			if (st.getObjectVar().getValue() == null)
-				freeVars.add(st.getObjectVar().getName());
-			return freeVars;
-		}
-
-		if (tupleExpr instanceof Projection) {
-			Projection p = (Projection) tupleExpr;
-			return new ArrayList<>(p.getBindingNames());
-		}
-
-		if (tupleExpr instanceof BindingSetAssignment) {
-			return new ArrayList<>();
-		}
-
-		if (tupleExpr instanceof Extension) {
-			// for a BIND extension in our cost model we work with 0 free vars
-			return new ArrayList<String>();
-		}
-
-		if (tupleExpr instanceof ArbitraryLengthPath) {
-			return getFreeVars(((ArbitraryLengthPath) tupleExpr).getPathExpression());
-		}
-
-		if (tupleExpr instanceof LeftJoin) {
-			// for our cost model we treat as expensive and work with 0 free vars
-			return new ArrayList<String>();
-		}
-
-		log.warn("Type " + tupleExpr.getClass().getSimpleName()
-				+ " not supported for cost estimation. If you run into this, please report a bug.");
-		return new ArrayList<String>();
 	}
 }
